+++ packages/mrm-preset-YOUR_SCOPE/typescript/index.js
@@ -52,6 +52,6 @@ module.exports.description = "Migrate typescript to global setting";
 module.exports.parameters = {
   configScope: {
     type: "input",
-    default: "@drmikecrowe",
+    default: "@YOUR_SCOPE",
   },
 };