+++ packages/stylelint-config/CHANGELOG.md
@@ -3,26 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## 0.1.1 (2021-07-10)
-
-**Note:** Version bump only for package @drmikecrowe/stylelint-config
+# [0.2.0](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/stylelint-config@0.1.3...@YOUR_SCOPE/stylelint-config@0.2.0) (2021-07-28)
 
 
+### Features
+
+* general package/doc cleanup ([c69a8f6](https://github.com/YOUR_SCOPE/configs/commit/c69a8f60a03531f44d7996955d48d522d9637427))
 
 
 
-## [0.1.2](https://github.com/drmikecrowe/configs/compare/@drmikecrowe/stylelint-config@0.1.1...@drmikecrowe/stylelint-config@0.1.2) (2021-01-30)
 
-**Note:** Version bump only for package @drmikecrowe/stylelint-config
 
-## [0.1.1](https://github.com/drmikecrowe/configs/compare/@drmikecrowe/stylelint-config@0.1.0...@drmikecrowe/stylelint-config@0.1.1) (2020-09-14)
+## [0.1.3](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/stylelint-config@0.1.1...@YOUR_SCOPE/stylelint-config@0.1.3) (2021-07-27)
 
 ### Bug Fixes
 
--   stylelint ignore files ([9cedec2](https://github.com/drmikecrowe/configs/commit/9cedec230186d4d1cdbb6a02188c1bd8baf4c00e))
+- fixing typescript config ([90388c4](https://github.com/YOUR_SCOPE/configs/commit/90388c4a744ba11070f668e752123d549994c4fb))
+
+## [0.1.2](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/stylelint-config@0.1.1...@YOUR_SCOPE/stylelint-config@0.1.2) (2021-07-27)
+
+### Bug Fixes
+
+- fixing typescript config ([90388c4](https://github.com/YOUR_SCOPE/configs/commit/90388c4a744ba11070f668e752123d549994c4fb))
+
+## 0.1.1 (2021-07-10)
+
+**Note:** Version bump only for package @YOUR_SCOPE/stylelint-config
+
+## [0.1.2](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/stylelint-config@0.1.1...@YOUR_SCOPE/stylelint-config@0.1.2) (2021-01-30)
+
+**Note:** Version bump only for package @YOUR_SCOPE/stylelint-config
+
+## [0.1.1](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/stylelint-config@0.1.0...@YOUR_SCOPE/stylelint-config@0.1.1) (2020-09-14)
+
+### Bug Fixes
+
+- stylelint ignore files ([9cedec2](https://github.com/YOUR_SCOPE/configs/commit/9cedec230186d4d1cdbb6a02188c1bd8baf4c00e))
 
 # 0.1.0 (2020-09-14)
 
 ### Features
 
--   stylelint-config ([2c2310e](https://github.com/drmikecrowe/configs/commit/2c2310efbdb36e9eb00b778f0eeb09054aa6fd1d))
+- stylelint-config ([2c2310e](https://github.com/YOUR_SCOPE/configs/commit/2c2310efbdb36e9eb00b778f0eeb09054aa6fd1d))