+++ packages/typescript/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/configs-typescript@0.1.2...@YOUR_SCOPE/configs-typescript@0.1.3) (2021-07-27)
+
+**Note:** Version bump only for package @YOUR_SCOPE/configs-typescript
+
+## [0.1.2](https://github.com/YOUR_SCOPE/configs/compare/@YOUR_SCOPE/configs-typescript@0.1.1...@YOUR_SCOPE/configs-typescript@0.1.2) (2021-07-27)
+
+**Note:** Version bump only for package @YOUR_SCOPE/configs-typescript
+
 ## 0.1.1 (2021-07-10)
 
-**Note:** Version bump only for package @drmikecrowe/configs-typescript
+**Note:** Version bump only for package @YOUR_SCOPE/configs-typescript